Added reverse_string.c
I created a function called reverse that takes a C-string as its argument and reverses the string in place. After the function is done, it returns the address of the first element of the C-string as a character pointer. If the string contains "Happy Birthday!" initially, calling the function will result in the string being reversed to "!yadhtriB yppaH."